Fix bugzilla bug 294225 : Excel report generates OutOfMemory error.

+++ engine/org.eclipse.birt.report.engine.emitter.prototype.excel/src/org/eclipse/birt/report/engine/emitter/excel/BlankData.java
@@ -79,7 +79,7 @@ public class BlankData extends Data
 		}
 	}
 
-	public double getHeight( )
+	public float getHeight( )
 	{
 		if ( data != null )
 		{
@@ -88,7 +88,7 @@ public class BlankData extends Data
 		return super.getHeight( );
 	}
 
-	public void setHeight( double height )
+	public void setHeight( float height )
 	{
 		if ( data != null )
 		{
